Move base target logic out of effect-application and into its own element
