+++ templates/chat/parts/card-face.hbs
@@ -58,6 +58,11 @@
     {{/if}}
     {{/each}}
 
+    {{!-- Targets --}}
+    {{#if showTargets}}
+    <recorded-targets></recorded-targets>
+    {{/if}}
+
     {{!-- Buttons --}}
     {{#if buttonGroups}}
     <section class="icon-row">
